Move EObjectNodePointer out of JavaXPathFactoryImpl
